rename file, delete unused code
